% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data123.R
\docType{data}
\name{data123}
\alias{data123}
\title{data123}
\format{
A data frame with 300 rows and 4 variables:
\describe{
   \item{X}{Time to event X.}
   \item{deltaX}{Event indicator for event X (1 - event; 0 - censoring event).}
   \item{Y}{Time to event Y}
   \item{deltaY}{Event indicator for event Y (1 - event; 0 - censoring event).}
}
}
\source{
Simulated data.
}
\usage{
data123
}
\description{
Bivariate survival data simulated from Frank's copula family with parameter 4.426265, which is equivalent to Spearman's correlation of 0.6. In the first 100 rows, the follow-up time for all observations is restricted to 2, and some observations may be censored before the follow-up time. In rows 101:200, the follow-up time is not restricted, but approximately thirty percent of observations are independently censored. In rows 201:300, there is no censoring.
}
\keyword{datasets}
