% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survSpearman.R
\name{survSpearman}
\alias{survSpearman}
\title{Computes Spearman's Correlation for Bivariate Survival Data.}
\usage{
survSpearman(X = NULL, Y = NULL, deltaX = NULL, deltaY = NULL, data = NULL,
tauX = Inf, tauY = Inf, bivarSurf = NULL)
}
\arguments{
\item{X}{Time to event or censoring for variable \code{X}. It indicates time to event if argument \code{deltaX}=1 and time to censoring if argument \code{deltaX}=0.}

\item{Y}{Time to event or censoring for variable \code{Y}. It indicates time to event if argument \code{deltaY}=1 and time to censoring if argument \code{deltaY}=0.}

\item{deltaX}{Event indicator for variable \code{X}. \code{deltaX=1} if the event is observed and \code{0} if it is censored.}

\item{deltaY}{Event indicator for variable \code{Y}. \code{deltaY=1} if the event is observed and \code{0} if it is censored.}

\item{data}{Data frame containing variables (arguments) \code{X}, \code{Y}, \code{deltaX}, and \code{deltaY}.}

\item{tauX}{The \code{X} value that defines the restricted region for \code{X} variable.}

\item{tauY}{The \code{Y} value that defines the restricted region for \code{Y} variable.}

\item{bivarSurf}{A matrix containing the marginal and joint survival probabilities. The first column is the marginal survival probability corresponding to variable \code{X}. The first row is the marginal survival probability corresponding to variable \code{Y}. The rest of the matrix contains the joint survival probabilities. The row names of \code{bivarSurf} are ordered \code{X}-values. The column names of \code{bivarSurf} are ordered \code{Y}-values. Element \code{bivarSurf[1,1]} equals 1. Its row and column name is \code{'0'} (see the documentation for the return value \code{DabrowskaEst} in function \code{survDabrowska}).}
}
\value{
The function returns the following list of values. \code{'Restricted region set by user'} is a character vector of two user-specified restricted region values, \code{tauX} and \code{tauY}. \code{'Effective restricted region'} is character vector of two values that define the effective restricted region, the values that are just above the latest observed event times within the user-specified restricted region. \code{'Correlation'} is a numeric vector of two correlation values: the highest rank Spearman's correlation (\code{'HighestRank'}) and the restricted region Spearman's correlation (\code{'Restricted'}), where the restricted region is defined by the values in \code{'Effective restricted region'}.
}
\description{
Computes non-parametric estimates of Spearman's rank correlation for bivariate survival data. Two correlations are returned: a highest rank correlation that can be interpreted as Spearman's correlation after assigning a highest rank to observations beyond a specified region, and a restricted correlation that estimates Spearman's correlation within the specified region.
}
\details{
The function computes the highest rank and restricted Spearman's correlations with bivariate survival data. The data can be supplied in three ways: 1) as vectors \code{X}, \code{Y}, \code{deltaX}, and \code{deltaY}; 2) as data frame \code{data} that contains the variables mentioned in 1); and 3) as matrix \code{bivarSurf} containing marginal and joint survival probabilities. If \code{bivarSurf} is not \code{NULL} then 1) and 2) are ignored. If \code{bivarSurf} is \code{NULL} and \code{data} is not then 2) is used. If \code{bivarSurf} and \code{data} are \code{NULL} then 1) is used. The highest rank correlation is the Spearman's correlation that can be interpreted as Spearman's rank correlation computed after assigning the highest rank to the events outside of \code{tauX} and \code{tauY}. The restricted Spearman's correlation is Spearman's correlation computed within the restricted region defined by \code{tauX} and \code{tauY}. Note that given \code{tauX} and \code{tauY} the survival probability is estimated using the values that are just above the latest observed event times within that region, what we call an effective restricted region. This means that if, for example, \code{tauX} is greater than the latest observed event time for \code{X} variable and \code{tauY} is greater than the latest observed event time for \code{Y} variable, then \code{tauX} and \code{tauY} do not affect the correlation values since the effective restricted region remains the same (as defined by the maximum observed event or censoring event times). The method of Dabrowska can result in negative probability mass for some points. This may result in zero or negative probability of failure in the restricted region, in which case the restricted Spearman's correlation cannot be computed and NA value is returned. This only happens when the sample size is small and censoring is heavy.
}
\examples{
### Compute correlation from data
X <- c(0.5, 0.6, 0.7, 0.8)
Y <- c(0.44, 0.77, 0.88, 0.99)
deltaX <- c(1, 0, 1, 1)
deltaY <- c(1, 1, 1, 1)
survSpearman(X, Y, deltaX, deltaY)
survSpearman(X, Y, deltaX, deltaY, tauX = 100, tauY = 100)
survSpearman(X, Y, deltaX, deltaY, tauX = 100, tauY = 0.99)
survSpearman(X, Y, deltaX, deltaY, tauX = 0.8, tauY = 0.99)

### Compute correlation from survival surface
someSurf <- survDabrowska(X, Y, deltaX, deltaY)$DabrowskaEst
survSpearman(tauY = 0.9, bivarSurf = someSurf)

}
\references{
Dabrowska, D. M. (1988) Kaplan–Meier estimate on the plane. The Annals of Statistics 16, 1475–1489.

Eden, S.K., Li, C., Shepherd B.E. (2021). Non-parametric Estimation of Spearman's Rank Correlation with Bivariate Survival Data. Biometrics (under revision).
}
\author{
Svetlana K Eden, \email{svetlanaeden@gmail.com}
}
\keyword{Spearman}
\keyword{bivariate}
\keyword{correlation}
\keyword{survival}
