% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unadjusted.CorPSRs.R
\name{unadjusted.CorPSRs}
\alias{unadjusted.CorPSRs}
\title{Unadjusted correlation of probability scale residuals (PSRs)}
\usage{
unadjusted.CorPSRs(X, Y, deltaX, deltaY)
}
\arguments{
\item{X}{Time to event or censoring for variable \code{X}. It indicates time to event if argument \code{deltaX}=1 and time to censoring if argument \code{deltaX}=0.}

\item{Y}{Time to event or censoring for variable \code{Y}. It indicates time to event if argument \code{deltaY}=1 and time to censoring if argument \code{deltaY}=0.}

\item{deltaX}{Event indicator for variable \code{X}. \code{deltaX=1} if the event is observed and \code{0} if it is censored.}

\item{deltaY}{Event indicator for variable \code{Y}. \code{deltaY=1} if the event is observed and \code{0} if it is censored.}
}
\value{
The point estimate, standard error, P-value, and confidence interval based on the \code{0.95\%} significance level.
}
\description{
Unadjusted correlation of probability scale residuals (PSRs).
}
\details{
Computes the unadjusted correlation of probability scale residuals. The variance is computed using the Kaplan-Meier estimating equations proposed by Stute [1995].
}
\examples{
X = c(0.5, 0.61, 0.6, 0.8, 0.78, 0.7, 0.9)
Y = c(0.44, 0.15, 0.77, 0.88, 0.22, 0.99, .33)
deltaX = c(1, 0, 1, 1, 0, 1, 0)
deltaY = c(1, 0, 1, 0, 1, 1, 1)
unadjusted.CorPSRs(X, Y, deltaX, deltaY)

}
\references{
Winfried Stute. The central limit theorem under random censorship. The Annals of Statistics, pages 422–439, 1995.

Svetlana K Eden, Chun Li, and Bryan E Shepherd. Spearman’s rank correlation adjusting for covariates in bivariate survival data. In preparation, 2020.
}
\author{
Svetlana K Eden, \email{svetlanaeden@gmail.com}
}
\keyword{probability}
\keyword{residuals}
\keyword{scale}
